nexus/requirements.txt
# NEXUS Trading System — Dependencies
# Python 3.13

# pip install -r nexus/requirements.txt


# Exchange connectivity
python-binance>=1.0.19
ccxt>=4.4.56
websockets>=14.1
iqoptionapi>=7.0.2

# Machine Learning & Deep Learning
tensorflow>=2.18.0
stable-baselines3>=2.4.1
gymnasium>=1.2.3
scikit-learn>=1.8.0

# LLM / NLP
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.1.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langchain-openai>=0.2.0

# Data & Analysis
pandas>=2.2.3
numpy>=2.2.1
ta>=0.11.0
scipy>=1.15.1
yfinance>=0.2.52

# Database
sqlalchemy>=2.0.36
asyncpg>=0.30.0

# Telegram Bot
python-telegram-bot>=21.9

# HTTP & Async
requests>=2.32.3
aiohttp>=3.11.11

# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.4
tenacity>=9.0.0
psutil>=6.1.1
tabulate>=0.9.0
PyYAML>=6.0.2

# Backtesting
backtrader2>=1.9.76.123
matplotlib>=3.10.0
quantstats>=0.0.62

# Scheduling
schedule>=1.2.2

# Monitoring & Dashboard
fastapi>=0.115.6
uvicorn>=0.34.0
python-multipart>=0.0.20
PyJWT>=2.9.0
redis>=5.2.1
prometheus-client>=0.21.1

# Development & Testing
pytest>=8.3.4
pyright>=1.1.391
pyre-check>=0.9.22
